+++ src/reuse/report.py
@@ -225,20 +225,26 @@ class ProjectReport:
         # Setup report data container
         data: dict[str, Any] = {
             "non_compliant": {
-                "bad_licenses": self.bad_licenses,
-                "deprecated_licenses": [
-                    str(file) for file in self.deprecated_licenses
-                ],
-                "licenses_without_extension": self.licenses_without_extension,
-                "missing_licenses": self.missing_licenses,
-                "unused_licenses": [str(file) for file in self.unused_licenses],
-                "read_errors": [str(file) for file in self.read_errors],
-                "missing_copyright_info": [
-                    str(file) for file in self.files_without_copyright
-                ],
-                "missing_licensing_info": [
-                    str(file) for file in self.files_without_licenses
-                ],
+                "bad_licenses": list(sorted(self.bad_licenses)),
+                "deprecated_licenses": list(
+                    sorted(str(file) for file in self.deprecated_licenses)
+                ),
+                "licenses_without_extension": list(
+                    sorted(self.licenses_without_extension)
+                ),
+                "missing_licenses": list(sorted(self.missing_licenses)),
+                "unused_licenses": list(
+                    sorted(str(file) for file in self.unused_licenses)
+                ),
+                "read_errors": list(
+                    sorted(str(file) for file in self.read_errors)
+                ),
+                "missing_copyright_info": list(
+                    sorted(str(file) for file in self.files_without_copyright)
+                ),
+                "missing_licensing_info": list(
+                    sorted(str(file) for file in self.files_without_licenses)
+                ),
             },
             "files": [],
             "summary": {
@@ -254,7 +260,7 @@ class ProjectReport:
         # Populate 'summary'
         number_of_files = len(self.file_reports)
         data["summary"] = {
-            "used_licenses": list(self.used_licenses),
+            "used_licenses": list(sorted(self.used_licenses)),
             "files_total": number_of_files,
             "files_with_copyright_info": number_of_files
             - len(self.files_without_copyright),